Refactor withLoading in own file
